[EasyWebhook] Make sure MultiplierWebhookRetryStrategy returns waiting time as integer

+++ packages/EasyWebhook/src/RetryStrategies/MultiplierWebhookRetryStrategy.php
@@ -51,7 +51,7 @@ class MultiplierWebhookRetryStrategy implements WebhookRetryStrategyInterface
 
     public function getWaitingTime(WebhookInterface $webhook): int
     {
-        $delay = $this->delayMilliseconds * \pow($this->multiplier, $webhook->getCurrentAttempt());
+        $delay = (int)($this->delayMilliseconds * \pow($this->multiplier, $webhook->getCurrentAttempt()));
 
         return $delay > $this->maxDelayMilliseconds && $this->maxDelayMilliseconds !== null
             ? $this->maxDelayMilliseconds
